operator: Move ConfigMap resource to own file.
